default page partials added.

+++ Core/CarBook.Application/Features/CQRS/Commands/CarCommands/CreateCarCommand.cs
@@ -14,8 +14,8 @@ namespace CarBook.Application.Features.CQRS.Commands.CarCommands
         public string CoverImageUrl { get; set; }
         public int Km { get; set; }
         public string Transmission { get; set; }
-        public byte Seat { get; set; }
-        public byte Luggage { get; set; }
+        public int Seat { get; set; }
+        public int Luggage { get; set; }
         public string Fuel { get; set; }
         public string BigImageUrl { get; set; }
     }
