+++ Core/Domain/Entities/Car.cs
@@ -15,8 +15,8 @@ namespace Domain.Entities
         public string CoverImageUrl { get; set; }
         public int Km { get; set; }
         public string Transmission { get; set; }
-        public byte Seat { get; set; }
-        public byte Luggage { get; set; }
+        public int Seat { get; set; }
+        public int Luggage { get; set; }
         public string Fuel { get; set; }
         public string BigImageUrl { get; set; }
         public List<CarFeature> CarFeatures { get; set; }
